+++ src/units/roles/miners/sk_miner.rs
@@ -50,8 +50,8 @@ impl Kind for SKMiner {
         MovementProfile::RoadsOneToTwo
     }
 
-    fn respawn_timeout(&self, creep: Option<&Creep>) -> Option<u32> {
-        creep.map(|c| c.body().len() as u32 * 3).or(Some(0))
+    fn respawn_timeout(&self, creep: Option<&Creep>) -> Option<usize> {
+        creep.map(|c| c.body().len() * 3).or(Some(0))
     }
 
     fn get_task(&self, _: &Creep, home: &mut Shelter) -> Task {
